Commodore64: Removed a lot of functions in the motherboard glue to simplify things.

+++ BizHawk.Emulation/Computers/Commodore64/MOS/Sid.cs
@@ -630,8 +630,8 @@ namespace BizHawk.Emulation.Computers.Commodore64.MOS
 		private bool filterSelectLoPass;
 		private bool filterSelectHiPass;
 		private int potCounter;
-		private byte potX;
-		private byte potY;
+		private int potX;
+        private int potY;
 		private int[] voiceOutput;
 		private Voice[] voices;
 		private int volume;
